Counters types are set as uint64_t

+++ include/server.h
@@ -29,7 +29,7 @@ public:
     bool init(int argc, const char** argv, ConfigOpts& configOpts);
     RunRes run();
 
-    void getThreadPoolInfo(int& activeWorkers, int& expelledWorkers) const;
+    void getThreadPoolInfo(uint64_t& activeWorkers, uint64_t& expelledWorkers) const;
 protected:
     virtual bool initConfigOption(int argc, const char** argv, ConfigOpts& configOpts);
     virtual void initMisc(ConfigOpts& configOpts);
